requirements.txt
aiogram==3.22.0
openai>=1.107.0
python-dotenv==1.1.1
aiosqlite==0.21.0
